Fix a bug which caused old events to be streamed live.
This would happen if you were on a HS then logged out and logged in to
another HS. The /events request from the older HS was still pending, which
would timeout and store the token from the old HS to be used on the new HS.
This token would be a valid token and so the new HS would return events from
that token. Add unit tests to check that /events requests are cancelled from
now on.